Add files via upload
At the moment it reaches the destination page (the hashtag you tell it to search for) but is not programmed to like pictures as of now. Unlike the original, this one can actually type in the search box and search for the term (hashtag); simulating a human searching for a tag. The functionality to like pictures is currently in the works.